pass_reverse/captcha_solver/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow
numpy>=1.26.0
tensorflow>=2.16.0
keras>=3.0.0

